Commit missing changes
